pyproject.toml
[tool.poetry]
name = "electrolux-group-developer-sdk"
version = "0.0.22"
description = "A Python async client library for interacting with the Electrolux Developer Portal API."
authors = ["Electrolux Developer Portal Team <devportal@electrolux.com>"]
readme = "README.md"
keywords = ["home", "automation", "electrolux", "aeg", "frigidaire", "+home"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.10"
aiohttp = ">=3.12.2,<4.0.0"
pytest = ">=8.3.5,<9.0.0"
pytest-asyncio = ">=1.0.0,<2.0.0"
pyjwt = ">=2.10.1,<3.0.0"
pydantic = ">=2.11.5,<3.0.0"
aioresponses = ">=0.7.8,<0.8.0"
freezegun = ">=1.4.0,<2.0.0"
tzdata = ">=2024.1,<2025.0"

[tool.poetry.group.dev.dependencies]
setuptools = "<81"
pytest-cov = "^6.1.1"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
